% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amax.R
\name{amax}
\alias{amax}
\title{Local Maxima}
\usage{
amax(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
returns the indicies of local maxima. If \code{x[1] = max}, then
        it is ignored.
}
\description{
Determines all local maxima from a real valued vector.
}
\details{
The purpose is to detect all local maxima in a real valued 1D vector.
If the first element \code{x[1]} is the global maximum, it is ignored,
because there is no information about the previous element. If there
is a plateau, the first edge is detected.
}
\examples{

a <- c(1,2,3,2,1,5,5,4)
amax(a) # 3, 6
}
