
\name{ar.sdf}
\alias{ar.sdf}
\title{Calculate the spectral density function (SDF) for an AR process}
\description{}
\usage{ ar.sdf(freqs, phi, sigma2=1, delta.t=1) }
\arguments{
\item{freqs}{ frequencies to evaluate the SDF at }
\item{phi}{ AR parameters }
\item{sigma2}{ innovation variance }
\item{delta.t}{ sampling interval }
}
\value{ vector of SDF values }
\details{}
\author{Peter F. Craigmile}
\keyword{}

