
\name{ar2.example}
\docType{data}
\alias{ar2.example}
\title{Coefficients for an example AR(2) process}
\description{}
\usage{ar2.example}
\format{A vector containing the coefficients of the AR(2) process.}
\source{Equation 45 of Percival and Walden (1993).}
\references{ Percival and Walden (1993) \emph{Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques}. Cambridge: Cambridge University Press. }
\keyword{datasets}
