
\name{ar4.example}
\docType{data}
\alias{ar4.example}
\title{Coefficients for an example AR(4) process}
\description{}
\usage{ar4.example}
\format{A vector containing the coefficients of the AR(4) process.}
\source{Equation 46a of Percival and Walden (1993).}
\references{ Percival and Walden (1993) \emph{Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques}. Cambridge: Cambridge University Press.}
\keyword{datasets}
