% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari.R
\name{ari}
\alias{ari}
\title{Anthocyanin Reflectance Index (ARI)}
\usage{
ari(green, red_edge_1)
}
\arguments{
\item{green}{Green spectral band data.}

\item{red_edge_1}{Red Edge 1 spectral band data.}
}
\value{
ARI data.
}
\description{
Calculate the Anthocyanin Reflectance Index (ARI).
}
\note{
This index can be used to vegetation-related applications.
}
\examples{

green <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
red_edge_1 <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

ari(green, red_edge_1)
}
