
\name{arma.sdf}
\alias{arma.sdf}
\title{Calculate the spectral density function (SDF) for an ARMA process}
\description{}
\usage{ arma.sdf(freqs, phi, theta, sigma2=1, delta.t=1) }
\arguments{
\item{freqs}{ frequencies to evaluate the SDF at }
\item{phi}{ AR parameters }
\item{theta}{ MA parameters }
\item{sigma2}{ innovation variance }
\item{delta.t}{ sampling interval }
}
\value{ vector of SDF values }
\details{}
\author{Peter F. Craigmile}
\keyword{}
