
\name{arma.var.from.sdf}
\alias{arma.var.from.sdf}
\title{Calculate the variance of an ARMA process using the spectral density function (SDF)}
\description{}
\usage{ arma.var.from.sdf(phi, theta, sigma2=1, delta.t=1) }
\arguments{
\item{phi}{ AR parameters (may be omitted) }
\item{theta}{ MA parameters (may be omitted) }
\item{sigma2}{ innovation variance }
\item{delta.t}{ sampling interval }
}
\value{ variance of the ARMA process }
\details{}
\author{Peter F. Craigmile}
\keyword{}
