% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndvi.R
\name{bndvi}
\alias{bndvi}
\title{Normalized Difference NIR/Blue Blue-normalized difference (BNDVI).}
\usage{
bndvi(blue, nir)
}
\arguments{
\item{blue}{Blue spectral band data.}

\item{nir}{NIR spectral band data.}
}
\value{
BNDVI data.
}
\description{
Calculate the BNDVI.
}
\note{
This index can be used to vegetation-related applications.
}
\examples{

blue <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
nir <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

bndvi(blue, nir)
}
