
\name{closest.Fourier.index}
\alias{closest.Fourier.index}
\title{Calculate closest Fourier index to a given frequency}
\description{}
\usage{closest.Fourier.index(freq, N, delta.t=1)}
\arguments{
  \item{freq}{frequency}
  \item{N}{length of times series}
  \item{delta.t}{sampling interval}
}
\value{}
\details{Calculates the closest Fourier index to 'freq', obtained
  when performing a spectral analysis on a time series of length 'N',
  with sampling interval 'delta.t'.}
\author{Peter F. Craigmile}
\keyword{}

