
\name{complex.chol}
\alias{complex.chol}
\title{Complex-valued Cholesky decomposition}
\description{}
\usage{complex.chol(x,upper=TRUE)}
\arguments{
   \item{x}{ Matrix to calculate the complex-valued Cholesky of }
   \item{upper}{ If true (the default), calculate the upper triangular
    form, if false calculate the lower triangular form. }
  }
\value{ A matrix containing the complex-valued Cholesky decomposition
  in either the upper or lower triangular form. }
\details{}
\references{}
\seealso{}
\examples{}
\author{Peter F. Craigmile}
\keyword{}
