
\name{cosine.taper}
\alias{cosine.taper}
\title{Calculate a 100p\% cosine taper}
\description{}
\usage{ cosine.taper(N, p) }
\arguments{
\item{N}{The length of the time series }
\item{p}{A single numeric value between 0 and 1, inclusive }
}
\value{ A vector of length N containing the cosine (100p)\% taper }
\references{ Percival and Walden (1993) \emph{Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques}. Cambridge: Cambridge University Press, Equation 209. }
\details{ }
\author{Peter F. Craigmile}
\keyword{}
