
\name{create.taper}
\alias{create.taper}
\title{Create a spectral taper of a given type}
\description{}
\usage{create.taper(taper.name, N, ...)}
\arguments{
\item{taper.name}{Name of the taper}
\item{N}{Length of the taper}
\item{...}{Extra arguments to pass to the taper}
}
\value{A taper as a vector or matrix (if a multiple taper)}
\details{ }
\author{Peter F. Craigmile}
\keyword{}
