
\name{Dirichlet.kernel}
\alias{Dirichlet.kernel}
\title{Calculate Dirichlet's kernel}
\description{}
\usage{Dirichlet.kernel(freqs, N)}
\arguments{
\item{freqs}{frequencies to evaluate at}
\item{N}{length of time series}
}
\value{Dirichlet's kernel at the evaluated frequencies}
\details{}
\references{Equation 1.3c of Percival and Walden (1993)}
\author{Peter F. Craigmile}
\keyword{}

