% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpdd.R
\name{dpdd}
\alias{dpdd}
\title{Dual-Pol Diagonal Distance (DPDD).}
\usage{
dpdd(vv, vh)
}
\arguments{
\item{vv}{VV band data.}

\item{vh}{VH band data.}
}
\value{
DPDD data.
}
\description{
Dual-Pol Diagonal Distance (DPDD).
}
\examples{

vv <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
vh <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

dpdd(vv, vh)
}
