% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi2}
\alias{evi2}
\title{Enhanced Vegetation Index 2 (EVI2).}
\usage{
evi2(red, nir)
}
\arguments{
\item{red}{Red spectral band data.}

\item{nir}{NIR spectral band data.}
}
\value{
EVI2 data.
}
\description{
Calculate the Enhanced Vegetation Index 2 (EVI2).
}
\note{
This index can be used to vegetation-related applications.
}
\examples{

red <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
nir <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

evi2(red, nir)
}
