
\name{Fejer.kernel}
\alias{Fejer.kernel}
\title{Calculate Fejer's kernel}
\description{}
\usage{ Fejer.kernel(freqs, N, delta.t=1)}
\arguments{
\item{freqs}{frequencies to evaluate at}
\item{N}{length of time series}
\item{delta.t}{sampling interval}
}
\value{Fejer's kernel at the evaluated frequencies}
\details{}
\references{Equation 198b of Percival and Walden (1993)}
\author{Peter F. Craigmile}
\keyword{}

