
\name{Fisher.test}
\alias{Fisher.test}
\title{Fisher's test for an unknown periodicity}
\description{}
\usage{Fisher.test(sp, alpha=.05)}
\arguments{
\item{sp}{ Spectral object to analyze. }
\item{alpha}{ Size of the test. }
}
\value{ hypothesis test ('h.test') object containing the results of Fisher's test for an unknown periodicity. }
\details{}
\references{ Percival and Walden (1993) \emph{Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques}. Cambridge: Cambridge University Press, Section 10.9. }
\examples{}
\author{Peter F. Craigmile}
\keyword{}
