
\name{Fourier.frequencies}
\alias{Fourier.frequencies}
\title{Calculate all the Fourier frequencies}
\description{}
\usage{Fourier.frequencies(N, delta.t=1) }
\arguments{
  \item{N}{length of times series}
  \item{delta.t}{sampling interval}
}
\value{}
\details{Calculates all the Fourier frequencies obtained when performing
  spectral analysis on a time series of length 'N',
  with sampling interval 'delta.tt'.  }
\author{Peter F. Craigmile}
\keyword{}

