% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{H}
\alias{H}
\title{The Hilbert-transformation}
\usage{
H(x)
}
\arguments{
\item{x}{real valued time series}
}
\value{
A numeric real valued vector is returned
}
\description{
The Hilbert-transform is a phase shifter, which represents the complex complement
to a real vauled signal. It is calculated in the complex frequency space of the
signal by using the Fourier transform. Finally, calculating \eqn{f = y + i*H(y)}
gives the analytic signal, with a one sided spectrum. (See \code{\link{analyticFunction}})
}
