\name{Hanning.taper}
\alias{Hanning.taper}
\title{Calculate a Hanning taper}
\description{}
\usage{ Hanning.taper(N) }
\arguments{
  \item{N}{ The length of the time series }
  }
\value{ A vector of length N containing the Hanning taper }
\references{ Percival and Walden (1993) \emph{Spectral Analysis for
    Physical Applic
    ations: Multitaper and Conventional Univariate
    Techniques}. Cambridge: Cambridge Univ
  ersity Press, Page 210. }
\details{ The Hanning taper is the same as a 100\% cosine taper}
\author{Peter F. Craigmile}
\keyword{}
