% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolationFFT.R
\name{interpolate.fft}
\alias{interpolate.fft}
\title{interpolates data using the Fourier back transform}
\usage{
interpolate.fft(y, x = NULL, n = NULL, xout = NULL)
}
\arguments{
\item{y}{numeric data vector to be interpolated}

\item{x}{numeric data vector with reference points}

\item{n}{number of new points}

\item{xout}{a vector new points}
}
\value{
A list with a \code{x} and \code{y} component is returned. The \code{e99}
value evaluates the error of the interpolation with respect to linear approximation
with the \code{approx()} function.
}
\description{
There are two ways to interpolate data from a given spectrum.
Frist, one can do zero padding to cover \code{n} new data points. Or, secound
the complex amplitude with the associated frequency is taken and evaluated
at given points \code{xout}. Doing that for all frequencies and amplitudes
will give the interpolation. The result is compared to linear approximation
for didactic reasons.
}
