% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specLOMB.R
\name{lmb}
\alias{lmb}
\title{Lomb-Scargle estimation function}
\usage{
lmb(f, dat, var_val)
}
\arguments{
\item{f}{frequency}

\item{dat}{spatial vector including locations and values}

\item{var_val}{variance of the data}
}
\description{
calculates the standard Lomb-Scargle estimation. The calculation is vectorized
to enhance calculation speed.
}
