
\name{ma.sdf}
\alias{ma.sdf}
\title{Calculate the spectral density function (SDF) for an MA process}
\description{}
\usage{ma.sdf(freqs, theta, sigma2=1, delta.t=1) }
\arguments{
\item{freqs}{frequencies to evaluate the SDF at}
\item{theta}{MA parameters}
\item{sigma2}{innovation variance}
\item{delta.t}{sampling interval}
}
\value{Vector of SDF values}
\details{}
\author{Peter F. Craigmile}
\keyword{}

