\name{mean.spect}
\alias{mean.spect}
\title{Spectral Mean}
\description{Average multiply tapered spectral estimates}
\usage{
	\method{mean}{spect}(x, \ldots)
}
\arguments{
  \item{x}{an item of class 'spect'}
  \item{...}{Extra arguments from the 'mean' function'}
}
\value{An object of class 'spect' which contains the averaged spectra.}
\details{Average the different spectra in the spect class 'sp'
           (normally used for multitapering).}
\references{}
\seealso{}
\examples{}
\author{Peter F. Craigmile}
\keyword{}
