
\name{multitaper.effective.bw}
\alias{multitaper.effective.bw}
\title{Calculate the effective bandwidth of a set of multitapers}
\description{}
\usage{ multitaper.effective.bw(tapers) }
\arguments{
\item{tapers}{ a matrix of multitapers }
}
\reference{ A. T. Walden, E. J. McCoy and D. B. Percival (1995), \emph{The Effective Bandwidth of a Multitaper Spectral Estimator}, Biometrika, 82, 201-14.}
\value{}
\details{}
\author{Peter F. Craigmile}
\keyword{}
