% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr.R
\name{nbr}
\alias{nbr}
\title{Normalized Burn Ratio (NBR).}
\usage{
nbr(nir, swir_2)
}
\arguments{
\item{nir}{NIR spectral band data.}

\item{swir_2}{SWIR 2 spectral band data.}
}
\value{
NBR data.
}
\description{
Normalized Burn Ratio (NBR).
}
\note{
This index can be used to burn-related applications.
}
\examples{

nir <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
swir_2 <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

nbr(nir, swir_2)
}
