% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr.R
\name{nbr2}
\alias{nbr2}
\title{Normalized Burn Ratio 2 (NBR2).}
\usage{
nbr2(swir_1, swir_2)
}
\arguments{
\item{swir_1}{SWIR 1 spectral band data.}

\item{swir_2}{SWIR 2 spectral band data.}
}
\value{
NBR2 data.
}
\description{
Normalized Burn Ratio 2 (NBR2).
}
\note{
This index can be used to burn-related applications.
}
\examples{

swir_1 <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
swir_2 <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

nbr2(swir_1, swir_2)
}
