% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi}
\alias{ndvi}
\title{Normalized Difference Vegetation Index (NDVI)}
\usage{
ndvi(red, nir)
}
\arguments{
\item{red}{Red spectral band data.}

\item{nir}{NIR spectral band data.}
}
\value{
NDVI data.
}
\description{
Calculate the Normalized Difference Vegetation Index (NDVI).
}
\note{
This index can be used to vegetation-related applications.
}
\examples{

red <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
nir <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

ndvi(red, nir)
}
