% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndwi.R
\name{ndwi}
\alias{ndwi}
\title{Normalized Difference Water Index (NDWI)}
\usage{
ndwi(green, nir)
}
\arguments{
\item{green}{Green spectral band data.}

\item{nir}{NIR spectral band data.}
}
\value{
NDWI data.
}
\description{
Calculate the Normalized Difference Water Index (NDWI).
}
\note{
This index can be used to water-related applications.
}
\examples{

green <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
nir <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

ndwi(green, nir)
}
