% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndwi.R
\name{ndwi2}
\alias{ndwi2}
\title{Normalized Difference Water Index (NDWI2)}
\usage{
ndwi2(green, nir)
}
\arguments{
\item{green}{Green spectral band data.}

\item{nir}{NIR spectral band data.}
}
\value{
NDWI2 data.
}
\description{
Calculate the Normalized Difference Water Index (NDWI2).
}
\note{
This index can be used to water-related applications.
}
\examples{

green <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5))
nir <- terra::rast(base::matrix(1:25, nrow = 5, ncol = 5)) + 1

ndwi2(green, nir)
}
