% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onDetach}
\alias{.onDetach}
\title{Reset multithread BLAS to default}
\usage{
.onDetach(libpath)
}
\arguments{
\item{libpath}{a character string giving the complete path to the package.}
}
\description{
This function is invoked automatically
}
