
\name{periodicity}
\alias{periodicity}
\title{Generate periodicity}
\description{}
\usage{periodicity(N, amps, freqs, phases, delta.t)}
\arguments{
\item{N}{ The length of the time series to generate. }
\item{amps}{ A vector of amplitudes. }
\item{freqs}{ A vector of frequencies. }
\item{phases}{ A vector of phases (default is to generate from Uniform[\eqn{-\pi},\eqn{\pi}])}
\item{delta.t}{ The sampling interval (The default is 1). }
}
\value{ The periodocity }
\details{}
\references{}
\examples{}
\author{Peter F. Craigmile}
\keyword{}
