
\name{periodogram}
\alias{periodogram}
\alias{pgram}
\title{Periodogram}
\description{Calculate periodogram}
\usage{periodogram(x, delta.t=deltat(x))}
\arguments{
  \item{x}{Time series object (the data)}
  \item{delta.t}{Sampling interval}
}
\value{ An object of class 'spect' containing the periodogram of 'x'. }
\details{Calculate the periodogram of a univariate time series of 'x'.}
\references{}
\seealso{}
\examples{}
\author{Peter F. Craigmile}
\keyword{}
