% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFFT.R
\name{print.fft}
\alias{print.fft}
\title{FFT-Plotting Function}
\usage{
\method{print}{fft}(x, ...)
}
\arguments{
\item{x}{lomb object}

\item{...}{not used}
}
\value{
This function returns nothing
}
\description{
It calls the summary function.
}
\examples{
# see summary.lomb() function
}
