% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLOMB.R
\name{print.lomb}
\alias{print.lomb}
\title{Lomb-Plotting Function}
\usage{
\method{print}{lomb}(x, ...)
}
\arguments{
\item{x}{lomb object}

\item{...}{not used}
}
\value{
This function returns nothing
}
\description{
It calls the summary function.
}
\examples{
# see summary.lomb() function
}
