
\name{rectangular.taper}
\alias{rectangular.taper}
\title{Calculate a rectangular taper}
\description{}
\usage{ rectangular.taper(N) }
\arguments{
  \item{N}{ The length of the time series }
  }
\value{ A vector of length N containing the rectangular taper }
\references{ Percival and Walden (1993) \emph{Spectral Analysis for
    Physical Applic
    ations: Multitaper and Conventional Univariate
    Techniques}. Cambridge: Cambridge Univ
  ersity Press, Page 209. }
\details{ The rectangular taper is the same a 0\% cosine taper}
\author{Peter F. Craigmile}
\keyword{}

