
\name{spect.max}
\alias{spect.max}
\title{The frequency of maximum spectrum}
\description{}
\usage{spect.max(sp, f1, f2)}
\arguments{
 \item{sp}{object of class 'spect'}
 \item{f1}{lower bound of the frequency}
 \item{f2}{upper bound of the frequency}
}
\value{ The frequency between 'f1' and 'f2' which has highest spectrum
 in the 'sp'. }
\details{}
\references{}
\seealso{}
\examples{}
\author{Peter F. Craigmile}
\keyword{}
