% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFFT.R
\name{summary.fft}
\alias{summary.fft}
\title{Summarize FFT objects}
\usage{
\method{summary}{fft}(object, p0 = 0.01, ...)
}
\arguments{
\item{object}{\code{lomb} object}

\item{p0}{False Alarm Probability (FAP) threshold, default 1\%}

\item{...}{not used}
}
\value{
a list of significant values of the spectral analysis
}
\description{
The function summarizes properties from the \code{class(fft)} object.
}
\details{
The false alarm probability threshold \code{p0} value can be
changed to modify the amount of significant peaks.
}
\examples{
# see spec.fft() example

}
