
\name{Thomson.crit}
\alias{Thomson.crit}
\title{Critical value of Thomson's F test}
\description{}
\usage{Thomson.crit(n.tapers, alpha)}
\arguments{
\item{n.tapers}{ The number of tapers }
\item{alpha}{ The alpha value to evaluate critical value }
}
\value{ alpha'th critical value }
\details{}
\references{ Percival and Walden (1993) \emph{Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques}. Cambridge: Cambridge University Press, Section 10.11. }
\examples{}
\author{Peter F. Craigmile}
\keyword{}
