
\name{transfer}
\alias{transfer}
\title{Calculate the transfer function}
\description{}
\usage{transfer(fs, filter)}
\arguments{
  \item{fs}{ a numerical vector of frequencies }
  \item{filter}{ a vector of filter coefficients }
  }
\value{ A numerical vector of transfer function values, the same length as 'fs'. }
\details{Calculate the transfer function for the 'filter' at the frequencies 'fs'.}
\references{}
\examples{}
\author{Peter F. Craigmile}
\keyword{}

