% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowfunctions.R
\name{win.cos}
\alias{win.cos}
\title{Cosine window function}
\usage{
win.cos(n)
}
\arguments{
\item{n}{data vector to be windowed}
}
\description{
This window function returns a vector of weights with means of a cosine window
}
\seealso{
\code{\link{Windowfunctions}}
}
