% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowfunctions.R
\name{win.hann}
\alias{win.hann}
\title{Hanning window function}
\usage{
win.hann(n, a = 2)
}
\arguments{
\item{n}{data vector to be windowed}

\item{a}{order of the window, default \code{a = 2}}
}
\description{
This window function returns a vector of weights with means of a
generlized Hann-window.
}
\seealso{
\code{\link{Windowfunctions}}
}
