% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowfunctions.R
\name{win.tukey}
\alias{win.tukey}
\title{Tukey window function}
\usage{
win.tukey(n, a = 0.5)
}
\arguments{
\item{n}{data vector to be windowed}

\item{a}{width of the rising and falling edge as ratio of the total data length}
}
\description{
This window function returns a vector of weights with means of a
Tukey-window. In contrast to a cosine window this function is more steep
at the beginning and the end. And it is 1 in the middle.
}
\seealso{
\code{\link{Windowfunctions}}
}
