% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmb}
\alias{lmb}
\title{Lomb-Scargle estimation function}
\usage{
lmb(x, y, omega)
}
\arguments{
\item{x}{spatial vector}

\item{y}{data vector}

\item{omega}{frequency vector}
}
\description{
calculates the Lomb-Scargle estimation
}
\details{
This wrapper function calls compiled C++ code, which does the job.
}

