
\name{multitaper.effective.bw}
\alias{multitaper.effective.bw}
\title{Calculate the effective bandwidth of a set of multitapers}
\description{}
\usage{ multitaper.effective.bw(tapers) }
\arguments{
\item{tapers}{ a matrix of multitapers }
}
\value{}
\details{}
\author{Peter F. Craigmile}
\keyword{}
