% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFFT.R
\name{plot.fft}
\alias{plot.fft}
\title{Plot \code{fft}-objects}
\usage{
\method{plot}{fft}(x, ...)
}
\arguments{
\item{x}{Object of the class \code{fft}}

\item{...}{further arguments to the plot functions}
}
\description{
Plot \code{fft}-objects
}
\examples{
# See spec.fft
}
\seealso{
\code{\link{spec.fft}}
}

